tests/wpt/web-platform-tests/navigation-api/currententrychange-event/navigation-navigate-same-doc.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));

  let oncurrententrychange_called = false;
  navigation.oncurrententrychange = t.step_func(e => {
    oncurrententrychange_called = true;
    assert_equals(e.from, navigation.entries()[0]);
    assert_equals(e.navigationType, "push");
    assert_equals(navigation.currentEntry.index, 1);
  });
  let result = navigation.navigate("#foo");
  assert_true(oncurrententrychange_called);
  await result.committed;
}, "currententrychange fires for navigation.navigate()");
</script>
